epde scripts update
